working on erf

+++ rems_backend/rems_api/views.py
@@ -2,8 +2,8 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.authentication import TokenAuthentication
-from rems_api.serializers import LocationSerializer, ArfSerializer
-from rems_api.models import Location, Arf
+from rems_api.serializers import LocationSerializer, ArfSerializer, ErfSerializer
+from rems_api.models import Location, Arf, Erf
 from rest_framework import status
 from django.http import Http404, HttpResponse, HttpResponseNotFound
 import os
@@ -67,6 +67,58 @@ class ArfDetail(APIView):
             module_dir = os.path.dirname(__file__)
             arf_path = os.path.join(module_dir, 'static/rems_api/')
             file_path = os.path.join(arf_path, arf.excel_sheet)
+            try:
+                os.remove(file_path)
+            except OSError:
+                pass
+        arf.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+class ErfList(APIView, TokenAuthentication):
+    permission_classes = (IsAuthenticated, )
+
+    def get(self, request):
+        # TODO: List file according to lodgin user
+        erfs = Erf.objects.all()
+        serializer = ErfSerializer(erfs, many=True)
+        return Response(serializer.data)
+
+    def post(self, request):
+        serializer = ErfSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class ErfDetail(APIView):
+    def get_object(self, pk):
+        try:
+            return Erf.objects.get(pk=pk)
+        except Erf.DoesNotExist:
+            raise Http404
+
+    def get(self, resquest, pk):
+        erf = self.get_object(pk)
+        serializer = ErfSerializer(erf)
+        return Response(serializer.data)
+
+    def put(self, request, pk):
+        erf = self.get_object(pk)
+        serializer = ErfSerializer(erf, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, pk):
+        erf = self.get_object(pk)
+        # TODO: Delte erf in database and file in server
+        if arf.excel_sheet:
+            module_dir = os.path.dirname(__file__)
+            erf_path = os.path.join(module_dir, 'static/rems_api/')
+            file_path = os.path.join(arf_path, erf.excel_sheet)
             try:
                 os.remove(file_path)
             except OSError:
